# Replace % inside loop of Random ctor