--- a/src/System.Runtime.Extensions/src/System/Random.cs
+++ b/src/System.Runtime.Extensions/src/System/Random.cs
@@ -76,7 +76,9 @@ namespace System
             {
                 for (int i = 1; i < 56; i++)
                 {
-                    SeedArray[i] -= SeedArray[1 + (i + 30) % 55];
+                    int n = i + 30;
+                    if (n >= 55) n -= 55;
+                    SeedArray[i] -= SeedArray[1 + n];
                     if (SeedArray[i] < 0) SeedArray[i] += MBIG;
                 }
             }