@chrisaut 2016-09-12
Second % operator in inner loop:

With this x64 also seems to be measurably faster for me:
500000 runs
new ~260ms x86
old ~570ms x86

new ~255ms x64
old ~380ms x64

```C#
            for (int j = 0; j < 15; j++)
            {
                sw.Restart();
                for (int i = 0; i < 500000, i++)
                {
                    new Random(10);
                }
                sw.Stop();
                Console.WriteLine(sw.ElapsedMilliseconds);
            }
```

Assembly before (x86):
```
                for (int i = 1; i < 56; i++)
028409DA  mov         edi,1  
028409DF  mov         eax,dword ptr [ebp-14h]  
028409E2  mov         esi,dword ptr [eax+4]  
                {
                    SeedArray[i] -= SeedArray[1 + (i + 30) % 55];
028409E5  mov         eax,dword ptr [ebp-14h]  
028409E8  mov         esi,dword ptr [eax+4]  
028409EB  cmp         edi,dword ptr [esi+4]  
028409EE  jae         02840A50  
028409F0  lea         ebx,[esi+edi*4+8]  
028409F4  lea         eax,[edi+1Eh]  
028409F7  mov         ecx,37h  
028409FC  cdq  
028409FD  idiv        eax,ecx  
028409FF  inc         edx  
02840A00  cmp         edx,dword ptr [esi+4]  
02840A03  jae         02840A50  
02840A05  mov         eax,dword ptr [esi+edx*4+8]  
02840A09  sub         dword ptr [ebx],eax  
                    if (SeedArray[i] < 0) SeedArray[i] += MBIG;
```
Assembly after (x86):
```
                for (int i = 1; i < 56; i++)
028406B5  mov         ecx,1  
028406BA  mov         edx,dword ptr [ebx+4]  
                {
                    int n = i + 30;
028406BD  lea         edi,[ecx+1Eh]  
                    if (n >= 55) n -= 55;
028406C0  cmp         edi,37h  
028406C3  jl          028406C8  
028406C5  add         edi,0FFFFFFC9h  
                    SeedArray[i] -= SeedArray[1 + n];
028406C8  mov         edx,dword ptr [ebx+4]  
028406CB  cmp         ecx,dword ptr [edx+4]  
028406CE  jae         0284071C  
028406D0  lea         esi,[edx+ecx*4+8]  
028406D4  inc         edi  
028406D5  cmp         edi,dword ptr [edx+4]  
028406D8  jae         0284071C  
028406DA  mov         eax,dword ptr [edx+edi*4+8]  
028406DE  sub         dword ptr [esi],eax  
                    if (SeedArray[i] < 0) SeedArray[i] += MBIG;
```
Assembly before (x64):
```
                for (int i = 1; i < 56; i++)
00007FFC9D2B0B3C  mov         r10d,1  
                {
                    SeedArray[i] -= SeedArray[1 + (i + 30) % 55];
00007FFC9D2B0B42  mov         rax,rcx  
00007FFC9D2B0B45  cmp         r10d,r8d  
00007FFC9D2B0B48  jae         00007FFC9D2B0BCC  
00007FFC9D2B0B4E  movsxd      rdx,r10d  
00007FFC9D2B0B51  lea         r11,[rax+rdx*4+10h]  
00007FFC9D2B0B56  mov         rdi,r11  
00007FFC9D2B0B59  mov         rbx,rcx  
00007FFC9D2B0B5C  lea         ebp,[r10+1Eh]  
00007FFC9D2B0B60  mov         eax,94F2095h  
00007FFC9D2B0B65  imul        ebp  
00007FFC9D2B0B67  mov         eax,edx  
00007FFC9D2B0B69  sar         eax,1  
00007FFC9D2B0B6B  mov         edx,eax  
00007FFC9D2B0B6D  shr         edx,1Fh  
00007FFC9D2B0B70  add         eax,edx  
00007FFC9D2B0B72  imul        eax,eax,37h  
00007FFC9D2B0B75  mov         edx,ebp  
00007FFC9D2B0B77  sub         edx,eax  
00007FFC9D2B0B79  lea         eax,[rdx+1]  
00007FFC9D2B0B7C  cmp         eax,r8d  
00007FFC9D2B0B7F  jae         00007FFC9D2B0BCC  
00007FFC9D2B0B81  movsxd      rax,eax  
00007FFC9D2B0B84  mov         eax,dword ptr [rbx+rax*4+10h]  
00007FFC9D2B0B88  sub         dword ptr [rdi],eax  
                    if (SeedArray[i] < 0) SeedArray[i] += MBIG;
```
Assembly after (x64):
```
                for (int i = 1; i < 56; i++)
00007FFC9D2D070A  mov         ecx,1  
                {
                    int n = i + 30;
00007FFC9D2D070F  lea         r9d,[rcx+1Eh]  
                    if (n >= 55) n -= 55;
00007FFC9D2D0713  cmp         r9d,37h  
00007FFC9D2D0717  jl          00007FFC9D2D071D  
00007FFC9D2D0719  add         r9d,0FFFFFFC9h  
                    SeedArray[i] -= SeedArray[1 + n];
00007FFC9D2D071D  mov         r10,rdx  
00007FFC9D2D0720  cmp         ecx,r8d  
00007FFC9D2D0723  jae         00007FFC9D2D0787  
00007FFC9D2D0725  movsxd      r11,ecx  
00007FFC9D2D0728  lea         r10,[r10+r11*4+10h]  
00007FFC9D2D072D  mov         r11,r10  
00007FFC9D2D0730  mov         rdi,rdx  
00007FFC9D2D0733  inc         r9d  
00007FFC9D2D0736  cmp         r9d,r8d  
00007FFC9D2D0739  jae         00007FFC9D2D0787  
00007FFC9D2D073B  movsxd      r9,r9d  
00007FFC9D2D073E  mov         r9d,dword ptr [rdi+r9*4+10h]  
00007FFC9D2D0743  sub         dword ptr [r11],r9d  
                    if (SeedArray[i] < 0) SeedArray[i] += MBIG;
```