@hughbe 2016-09-12
Test Innerloop Ubuntu14.04 Debug Build and Test