@stephentoub 2016-09-12
LGTM

Note, though, that this implementation is currently only used when running on .NET Native.  The copy of the source in the coreclr repo is used when running on coreclr.  If we're going to make such a change in one place, ideally we'd make it in both.

That said, I'm not sure why we don't just always use the corefx implementation.  @mellinoe, @weshaggard , do you know if there's any reason we can't just use this corefx implementation in System.Runtime.Extensions rather than type-forwarding on coreclr?  I don't see Random exposed out of any other mscorlib surface area that would impact type identity.