@chrisaut 2016-09-12
That is unfortunate, if the corefx version cannot be used, I'll prepare a pull request for the coreclr version. But I'll wait until you guys decide what to do.