@mellinoe 2016-09-12
> I don't see Random exposed out of any other mscorlib surface area that would impact type identity

Yeah, I would be surprised if there's anything like that. This is most likely just something that we never bothered to unify since it has been relatively untouched for a while. Unless there's something I'm not thinking of, we should be able to use the version here everywhere.