@stephentoub 2016-09-12
> Unless there's something I'm not thinking of, we should be able to use the version here everywhere.

Ok, I'll submit a PR to do that.