@stephentoub 2016-09-12
https://github.com/dotnet/corefx/pull/11653