@stephentoub 2016-09-13
@dotnet-bot test this please (to run recently added Random test)