@stephentoub 2016-09-13
@chrisaut, I'm going to close this PR: per the discussion in #11653 and https://github.com/dotnet/corefx/issues/11665, we're actually going to delete this code from corefx, as both coreclr and corert have their own public versions that'll be used (as I mentioned, the coreclr is already being used for things other than .NET Native, so merging this PR wouldn't help that).  Thanks for the interest in contributing, and please feel free to submit PRs to those repos for this change.